Add module for styles
