+++ gui/qt/lite_window.py
@@ -33,55 +33,12 @@ from chainkey.version import ELECTRUM_VERSION as electrum_version
 from chainkey.util import format_satoshis, age
 
 from main_window import ElectrumWindow
+from style import *
 import shutil
 
 from util import *
 
 bitcoin = lambda v: v * 100000000
-
-def IconButton(filename, parent=None):
-    pixmap = QPixmap(filename)
-    icon = QIcon(pixmap)
-    return QPushButton(icon, "", parent)
-
-
-def resize_line_edit_width(line_edit, text_input):
-    metrics = QFontMetrics(qApp.font())
-    # Create an extra character to add some space on the end
-    text_input += "A"
-    line_edit.setMinimumWidth(metrics.width(text_input))
-
-def load_theme_name(theme_path):
-    try:
-        with open(os.path.join(theme_path, "name.cfg")) as name_cfg_file:
-            return name_cfg_file.read().rstrip("\n").strip()
-    except IOError:
-        return None
-
-
-def theme_dirs_from_prefix(prefix):
-    if not os.path.exists(prefix):
-        return []
-    theme_paths = {}
-    for potential_theme in os.listdir(prefix):
-        theme_full_path = os.path.join(prefix, potential_theme)
-        theme_css = os.path.join(theme_full_path, "style.css")
-        if not os.path.exists(theme_css):
-            continue
-        theme_name = load_theme_name(theme_full_path)
-        if theme_name is None:
-            continue
-        theme_paths[theme_name] = prefix, potential_theme
-    return theme_paths
-
-def load_theme_paths():
-    theme_paths = {}
-    lite_theme_dir = os.path.join(util.data_dir(), 'lite_gui')
-    theme_paths.update(theme_dirs_from_prefix(lite_theme_dir))
-    return theme_paths
-
-
-
 
 class TransactionWindow(QDialog):
 
@@ -599,42 +556,9 @@ class ReceivePopup(QDialog):
         QCursor.setPos(center_mouse_pos)
         self.show()
 
-class MiniActuator:
+class MiniActuator(Actuator):
     """Initialize the definitions relating to themes and
     sending/receiving bitcoins."""
-
-
-    def __init__(self, main_window):
-        """Retrieve the gui theme used in previous session."""
-        self.g = main_window
-        self.theme_name = self.g.config.get('litegui_theme','Cleanlook')
-        self.themes = load_theme_paths()
-        self.load_theme()
-
-    def load_theme(self):
-        """Load theme retrieved from wallet file."""
-        try:
-            theme_prefix, theme_path = self.themes[self.theme_name]
-        except KeyError:
-            util.print_error("Theme not found!", self.theme_name)
-            return
-        full_theme_path = "%s/%s/style.css" % (theme_prefix, theme_path)
-        with open(full_theme_path) as style_file:
-            qApp.setStyleSheet(style_file.read())
-
-    def theme_names(self):
-        """Sort themes."""
-        return sorted(self.themes.keys())
-
-    def selected_theme(self):
-        """Select theme."""
-        return self.theme_name
-
-    def change_theme(self, theme_name):
-        """Change theme."""
-        self.theme_name = theme_name
-        self.g.config.set_key('litegui_theme',theme_name)
-        self.load_theme()
 
     def set_configured_exchange(self, set_exchange):
         use_exchange = self.g.config.get('use_exchange')
